requirements.txt
# ChartM3 Dependencies

# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
requests>=2.32.4

# Image processing
Pillow>=12.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
kaleido>=0.2.1

# Optional: for tokenization analysis
transformers>=4.53.0

# Optional: for VLM image processing
# qwen-vl-utils>=0.0.2

# Development dependencies (optional)
# pytest>=7.0.0
# black>=23.0.0
# isort>=5.12.0
